backend/requirements.txt
astroid==3.2.4
asttokens==2.4.1
black==24.8.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
debugpy==1.8.1
decorator==5.1.1
dill==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.4
Flask==3.0.3
fsspec==2024.6.1
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
ipykernel==6.29.4
ipython==8.24.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.43.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
more-itertools==10.3.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
numba==0.60.0
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
openai-whisper==20231117
packaging==24.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.26.1
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.2.6
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==5.4.1
pyzmq==26.0.3
regex==2024.5.15
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.0
systemd-python==234
tiktoken==0.7.0
tomli==2.0.1
tomlkit==0.13.2
torch==2.3.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.2
wadllib==1.3.6
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==1.0.0
